Plumbing border box through to viz, and adding clip rect to backdrop-filter
With this CL, the border box of the element containing backdrop-filter
is plumbed through to viz, so that the backdrop-filter effect can be
properly clipped at the border box of the element itself. This will keep
it from "bleeding" outside to the border rect of the entire layer (including
children).

This should fix most of the reported bugs related to the filtered effect
extending beyond the proper bounds.

Bug: 497522, 813796, 767997, 659501, 632979, 618913, 593307
Change-Id: I5f24ac79053b3779cea128aacabd60f9d66702f4